+++ packages/video_player_avplay/tizen/src/plus_player.cc
@@ -266,14 +266,14 @@ bool PlusPlayer::SetLooping(bool is_looping) {
 }
 
 bool PlusPlayer::SetVolume(double volume) {
-  LOG_INFO("[PlusPlayer] Volume: %f", volume);
-
   if (GetState(player_) < plusplayer::State::kPlaying) {
     LOG_ERROR("[PlusPlayer] Player is in invalid state");
     return false;
   }
-
-  if (!::SetVolume(player_, volume)) {
+  // dart api volume range[0,1], plusplaer volume range[0,100]
+  int new_volume = volume * 100;
+  LOG_INFO("[PlusPlayer] Volume: %d", new_volume);
+  if (!::SetVolume(player_, new_volume)) {
     LOG_ERROR("[PlusPlayer] Fail to set volume.");
     return false;
   }