+++ packages/video_player_avplay/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Check httper_headers pointer when creating media player.
 * Fix state check in SetVolume.
+* Add a case that plays the video of assets to the example.
 
 ## 0.3.2